requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
coverage==6.5.0
Django==3.2.25
django-cas-ng==5.0.1
django-dbconn-retry==0.1.7
django-dramatiq==0.11.6
django-redis==5.4.0
djangorestframework==3.14.0
dramatiq==1.16.0
entrypoints==0.4
Envelopes==0.4
flake8==7.0.0
flake8-coding==1.3.2
flake8-quotes==3.3.2
gunicorn==21.2.0
idna==3.6
jsonfield==3.1.0
lxml==5.2.1
mccabe==0.7.0
otpauth==1.0.1
packaging==24.0
pillow==10.2.0
prometheus_client==0.20.0
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pypng==0.20220715.0
python-cas==1.6.0
python-dateutil==2.8.2
pytz==2024.1
qrcode==7.4.2
raven==6.10.0
redis==5.0.3
requests==2.31.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
wheel==0.43.0
XlsxWriter==3.1.9
